backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aniso8601==7.0.0          # via graphene
attrs==19.3.0             # via pytest
beautifulsoup4==4.7.1     # via -r requirements.in
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
django-cors-headers==3.3.0  # via -r requirements.in
django-dbbackup==3.3.0    # via -r requirements.in
django-enumfields==0.10.1  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-extensions==2.0.7  # via -r requirements.in
django-polymorphic==2.0.2  # via -r requirements.in
django-tagulous==0.13.2   # via -r requirements.in
django-templated-mail==1.1.1  # via djoser
django-webpack-loader==0.6.0  # via -r requirements.in
django==2.2.6             # via -r requirements.in, django-cors-headers, django-dbbackup, django-polymorphic, graphene-django
djangorestframework==3.8.2  # via -r requirements.in
djoser==1.7.0             # via -r requirements.in
gitdb2==4.0.2             # via gitpython
gitdb==4.0.5              # via gitdb2
gitpython==2.1.10         # via -r requirements.in
graphene-django-optimizer==0.6.2  # via -r requirements.in
graphene-django==2.10.1   # via -r requirements.in
graphene==2.1.8           # via graphene-django
graphql-core==2.3.2       # via graphene, graphene-django, graphql-relay
graphql-relay==2.0.1      # via graphene
gunicorn==19.8.1          # via -r requirements.in
idna==2.6                 # via requests
more-itertools==8.3.0     # via pytest
packaging==20.4           # via pytest
pluggy==0.13.1            # via pytest
promise==2.3              # via graphene-django, graphql-core, graphql-relay
psycopg2==2.8.5           # via -r requirements.in
py==1.8.1                 # via pytest
pyparsing==2.4.7          # via packaging
pytest-django==3.5.1      # via -r requirements.in
pytest-html==2.1.1        # via -r requirements.in
pytest-metadata==1.9.0    # via pytest-html
pytest==5.4.3             # via pytest-django, pytest-html, pytest-metadata
python-memcached==1.59    # via -r requirements.in
pytz==2018.4              # via -r requirements.in, django, django-dbbackup
pyyaml==3.12              # via -r requirements.in
requests==2.18.4          # via -r requirements.in
rx==1.6.1                 # via graphql-core
singledispatch==3.4.0.3   # via graphene-django
six==1.15.0               # via django-dbbackup, django-extensions, graphene, graphene-django, graphql-core, graphql-relay, packaging, promise, python-memcached, singledispatch
smmap==3.0.4              # via gitdb
soupsieve==2.0.1          # via beautifulsoup4
sqlparse==0.3.1           # via django
urllib3==1.22             # via requests
wcwidth==0.2.3            # via pytest
